Rename module folders to lowercase
