+++ src/models/Stores.typeDef.js
